+++ backend/src/main.ts
@@ -54,5 +54,6 @@ async function bootstrap() {
 
   const port = process.env.PORT || 3000;
   await app.listen(port, "0.0.0.0");
+  app.enableShutdownHooks();
 }
 bootstrap();